feat!: controllers refactored (#57)
* fix pid typo

* pid refactore and unnecessary files removed

* fix example

* fix windows build for examples

* fix windows build

* fix windows build

* fix build

* fix windows build

+++ examples/analysis/FastFourierTransformRadix2/FastFourierTransformRadix2.cpp
@@ -20,7 +20,7 @@ namespace
 
             for (std::size_t k = 0; k < Length; ++k)
             {
-                float angle = -M_PI * static_cast<float>(k) / static_cast<float>(Length);
+                float angle = static_cast<float>(-M_PI) * static_cast<float>(k) / static_cast<float>(Length);
                 factors.emplace_back(std::cos(angle), std::sin(angle));
             }
         }
